init: start task remote-control
