+++ .eslintrc
@@ -1,0 +1,30 @@
+{
+  "plugins": ["prettier", "import", "@typescript-eslint"],
+  "extends": [
+      "plugin:prettier/recommended",
+      "prettier",
+      "plugin:@typescript-eslint/recommended"
+  ],
+  "parserOptions": {
+      "ecmaVersion": 2020,
+      "sourceType": "module"
+  },
+  "env": {
+      "es6": true,
+      "browser": true,
+      "node": true
+  },
+  "rules": {
+      "no-debugger": "off",
+      "no-console": "off",
+      "class-methods-use-this": "off",
+      "@typescript-eslint/no-explicit-any": "error",
+      "@typescript-eslint/no-var-requires": 0,
+      "prettier/prettier": [
+        "error", 
+      { "endOfLine": "auto" 
+      }
+    ]
+  },
+  "parser": "@typescript-eslint/parser"
+}
